Refactor for getDescription and package style
Update package style to accord with the example of
``packages/telescope-blank``.  Add i18n.  Simplify description setting
into a single setting for both meta descriptions and og:description. Use
``getTitle`` and ``getDescription``.
