+++ packages/telescope-seo/i18n/en.i18n.json
@@ -1,0 +1,6 @@
+{
+  "search_engine_optimization": "search engine optimization",
+  "seoDescription": "meta and og description",
+  "seoOgImage": "og:image",
+  "seoGenerateSitemap": "Generate Sitemap (requires restart)"
+}
