+++ packages/telescope-seo/package-tap.i18n
@@ -1,0 +1,5 @@
+{
+  "translation_function_name": "__",
+  "helper_name": "_",
+  "namespace": "project"
+}
