+++ packages/telescope-seo/package.js
@@ -1,24 +1,32 @@
 Package.describe({
   name: "telescope-seo",
   summary: "SEO extensions for Telescope",
-  version: "0.0.3"
+  version: "0.0.4"
 });
 
 Package.onUse(function(api) {
-
   api.use([
+    "templating",
     "underscore",
     "aldeed:simple-schema",
+    "tap:i18n",
     "iron:router",
     "telescope-lib",
     "telescope-base",
-    "telescope-tags",
+    "telescope-i18n",
     "manuelschoebel:ms-seo@0.4.1",
     "gadicohen:sitemaps@0.0.20"
   ]);
 
-  api.export([
-  ]);
+  // both
+  api.addFiles([
+    "lib/routes.js",
+    "lib/seo.js",
+    "package-tap.i18n"
+  ], ['client', 'server']);
 
-  api.addFiles("seo.js", ['client', 'server']);
+  // server
+  api.addFiles([
+    "lib/server/sitemaps.js"
+  ], ["server"]);
 });
