+++ packages/telescope-seo/versions.json
@@ -205,10 +205,6 @@
       "0.2.9"
     ],
     [
-      "telescope-tags",
-      "0.0.0"
-    ],
-    [
       "templating",
       "1.0.9"
     ],
